Q1-readiness rigor fixes: nested CV, significance, leakage & CI fixes
Audit-driven fixes so the cancelable ECG+PPG (BioHashing + IoM) experiments
meet Q1 evaluation standards.

ML methodology
- Nested CV hyperparameter tuning (--tune): group-aware GridSearchCV inner
  loop per outer fold; build_param_grid centres on the former fixed configs,
  removing selection-on-evaluation-data bias.
- Significance + multiple comparisons (--significance): new mwf/significance.py
  runs paired Nadeau-Bengio t-tests (previously defined but never called) with
  Benjamini-Hochberg FDR control -> significance.csv.
- Stolen-token z-norm de-biased: statistics now come from a disjoint impostor
  cohort instead of self-normalising the scored impostors (EER was deflated).
- Standardisation leakage: build_templates returns/accepts a FeatureScaler;
  holdout fits it on train only (leakage-free), via TemplateBundle.scaler.
- subject_holdout wired as unseen-subject verification (subject_holdout.csv),
  with robustness guards for small held-out cohorts.

Statistical reporting
- Bootstrap CIs added: stolen-token EER, cancelability d_sys/diversity,
  non-invertibility leakage gap, verification EER (Nadeau-Bengio).
- Diversity pass/fail referenced to an analytical chance |r| baseline instead
  of magic thresholds; key-level bootstrap over random keys.
- balanced_accuracy added to the reported metrics.

Robustness / reproducibility
- Non-invertibility claim scoped to the linear/min-norm attack family.
- Independent random base tokens for key sensitivity (i.i.d. draws).
- Explicit little-endian keystream seeding (cross-platform determinism).
- Per-run provenance manifest (git SHA, args, dep versions, uv.lock hash).
- README: single-session MIMIC limitation + claim scoping documented.

Tests: 120 passing (added test_significance.py + nested-CV/scaler/z-norm cases).

+++ mwf/keystream.py
@@ -46,7 +46,11 @@ def keystream_rng(token: str) -> np.random.Generator:
     if not token:
         raise ValueError("Token must be a non-empty string.")
     digest = hashlib.sha256(token.encode("utf-8")).digest()
-    seed_words = np.frombuffer(digest, dtype=np.uint32)
+    # Explicit little-endian word order ('<u4') so the seed — and hence every
+    # projection matrix — is bit-for-bit identical on big- and little-endian
+    # hardware. Native byte order ('uint32') would silently diverge across
+    # architectures and break the reproducibility claim.
+    seed_words = np.frombuffer(digest, dtype="<u4")
     return np.random.default_rng(np.random.SeedSequence(seed_words.tolist()))
